avformat/rtsp: Fix timeout option
92c40ef882be115e72d2aa02f9032b7ce88f8537 added a listen_timeout option
for sdp. This allowed a user to set variable timeout which was
originally hard coded to 10 seconds.

The commit used the initial_timeout variable to store the value. But
this variable is shared with rtsp where it's used to infer a "listen"
mode. Thus, the timeout value could not be set in rtsp, and the default
value (initial_timeout = -1) would give 100ms timeout.

This was attempted to be fixed in c8101aabee654f6d147a4d89f77fa73e18908610,
which changed the meaning of initial_timeout = -1 to be an infinite
timeout. However, it did not address the issue that the timeout could
still not be set. Being able to set the timeout is useful because it
allows to automatically reconfigure from a udp to tcp connection in the
lower transport.

In this commit this is fixed by using the stimeout variable to
store the timeout value.

+++ libavformat/rtsp.h
@@ -404,7 +404,7 @@ typedef struct RTSPState {
     /**
      * timeout of socket i/o operations.
      */
-    int stimeout;
+    int64_t stimeout;
 
     /**
      * Size of RTP packet reordering queue.
